Use Tabs on Assets details page
