+++ packages/html-vue/src/pages/Assets/Details.vue
@@ -9,169 +9,22 @@
 		<hr class="mt-4 mb-6 border-gray-100">
 
 		<div class="flex flex-col mb-4">
-			<h3 class="mb-4 text-xl font-bold">Asset details</h3>
+			<Tabs
+				v-model="activeTab"
+				:options="TABS"
+			>
+				<template #details>
+					<Details :name="name" />
+				</template>
 
-			<div class="rounded-lg border border-gray-200 overflow-hidden shadow-xs">
-				<table class="table-fixed w-full text-sm text-left">
-					<colgroup>
-						<col style="width: 210px">
-						<col style="width: 100%">
-					</colgroup>
+				<template #dependencies>
+					<Dependencies :name="name" :inputs="inputs" />
+				</template>
 
-					<tbody class="text-gray-500">
-						<tr>
-							<td class="p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">Path</td>
-							<td class="p-3 font-normal">{{ name }}</td>
-						</tr>
-
-						<tr class="border-t border-gray-100">
-							<td class="p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">File type</td>
-							<td class="p-3 font-normal capitalize">{{ asset.type }}</td>
-						</tr>
-
-						<tr
-							v-if="entrypoints.length"
-							class="border-t border-gray-100"
-						>
-							<td class="p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">Entrypoint(s)</td>
-							<td class="p-3 font-normal">
-								<ul>
-									<li
-										v-for="entrypoint in entrypoints"
-										:key="entrypoint"
-									>
-										<a
-											:href="router.getUrl( 'inputs/details', { item: entrypoint } )"
-											class="py-1 text-sm font-medium underline-offset-2 rounded-lg outline-hidden focus:ring focus:ring-gray-500 focus:border-gray-500 hover:underline"
-										>
-											{{ entrypoint }}
-										</a>
-									</li>
-								</ul>
-							</td>
-						</tr>
-
-						<!-- Uncompressed -->
-						<tr class="border-t border-gray-100">
-							<td class="p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">Original file size</td>
-							<td class="p-3 font-normal">{{ formatSize( asset.uncompressed ) }}</td>
-						</tr>
-						<tr class="border-t border-gray-100">
-							<td class="flex items-center justify-between p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">
-								Original download time
-
-								<span class="inline-flex" data-hover="Estimated download time on a slow 3G connection">
-									<IconInfo
-										:size="20"
-										class="ml-2 inline pointer-events-none"
-									/>
-								</span>
-							</td>
-							<td class="p-3 font-normal">{{ formatTime( downloadTimeOriginal ) }}</td>
-						</tr>
-
-						<!-- GZIP -->
-						<template v-if="asset.gzip">
-							<tr class="border-t border-gray-100">
-								<td class="p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">GZIP size</td>
-								<td class="p-3 font-normal">{{ formatSize( asset.gzip ) }}</td>
-							</tr>
-							<tr class="border-t border-gray-100">
-								<td class="flex items-center justify-between p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">
-									GZIP download time
-
-									<span class="inline-flex" data-hover="Estimated download time on a slow 3G connection">
-										<IconInfo
-											:size="20"
-											class="ml-2 inline pointer-events-none"
-										/>
-									</span>
-								</td>
-								<td class="p-3 font-normal">{{ formatTime( downloadTimeGzip ) }}</td>
-							</tr>
-						</template>
-
-						<!-- Brotli -->
-						<template v-if="asset.brotli">
-							<tr class="border-t border-gray-100">
-								<td class="p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">Brotli size</td>
-								<td class="p-3 font-normal">{{ formatSize( asset.brotli ) }}</td>
-							</tr>
-							<tr class="border-t border-gray-100">
-								<td class="flex items-center justify-between p-3 font-bold whitespace-nowrap bg-gray-50 border-r border-r-gray-100">
-									Brotli download time
-
-									<span class="inline-flex" data-hover="Estimated download time on a slow 3G connection">
-										<IconInfo
-											:size="20"
-											class="ml-2 inline pointer-events-none"
-										/>
-									</span>
-								</td>
-								<td class="p-3 font-normal">{{ formatTime( downloadTimeBrotli ) }}</td>
-							</tr>
-						</template>
-					</tbody>
-				</table>
-			</div>
-		</div>
-
-		<div class="flex flex-col mt-16 mb-4">
-			<h3 class="mb-4 text-xl font-bold">Dependencies ({{ dependencies.length }})</h3>
-
-			<div class="rounded-lg border border-gray-200 overflow-scroll max-h-[475px] shadow-xs">
-				<table class="table-fixed w-full text-sm text-left">
-					<colgroup>
-					</colgroup>
-
-					<tbody class="text-gray-700">
-						<tr
-							v-for="( dependency, index ) in dependencies"
-							:key="dependency"
-							class="[&:not(:first-child)]:border-t border-gray-100"
-						>
-							<td class="p-3 font-normal">
-								<span class="select-none mr-2">{{ index + 1 }}.</span>
-								<a
-									:href="router.getUrl( 'dependencies', { search: dependency, active: dependency } )"
-									class="px-2 py-1 text-sm font-medium underline-offset-2 rounded-lg outline-hidden focus:ring focus:ring-gray-500 focus:border-gray-500 hover:underline"
-								>
-									{{ dependency }}
-								</a>
-							</td>
-						</tr>
-					</tbody>
-				</table>
-			</div>
-		</div>
-
-		<div class="flex flex-col mt-16 mb-4">
-			<h3 class="mb-4 text-xl font-bold">Inputs ({{ inputs.length }})</h3>
-
-			<div class="rounded-lg border border-gray-200 overflow-scroll max-h-[475px] shadow-xs">
-				<table class="table-fixed w-full text-sm text-left">
-					<colgroup>
-					</colgroup>
-
-					<tbody class="text-gray-700">
-						<tr
-							v-for="( input, index ) in inputs"
-							:key="input"
-							class="[&:not(:first-child)]:border-t border-gray-100"
-						>
-							<td class="p-3 font-normal">
-								<span class="select-none mr-2">{{ index + 1 }}.</span>
-								<a
-									:href="router.getUrl( 'inputs/details', { item: input } )"
-									class="px-2 py-1 text-sm font-medium underline-offset-2 rounded-lg outline-hidden focus:ring focus:ring-gray-500 focus:border-gray-500 hover:underline"
-								>
-									{{ input }}
-								</a>
-							</td>
-						</tr>
-					</tbody>
-				</table>
-			</div>
+				<template #inputs>
+					<Inputs :name="name" :inputs="inputs" />
+				</template>
+			</Tabs>
 		</div>
 	</div>
 </template>
@@ -179,25 +32,22 @@
 <script setup lang="ts">
 import { computed } from 'vue';
 import { router } from '@/router.js';
-import { getAssetResource, getChunks, report } from '@/report.js';
-import { formatPath, formatSize, formatTime } from '@/format.js';
-import IconInfo from '@icon/Info.vue';
+import { getChunks } from '@/report.js';
+import { formatPath } from '@/format.js';
+import Tabs from '@/components/common/Tabs.vue';
+import Details from '@/components/assets/Details.vue';
+import Dependencies from '@/components/assets/Dependencies.vue';
+import Inputs from '@/components/assets/Inputs.vue';
 
-const SLOW_3G = 50 * 1024 * 8;
+const TABS = [
+	{ label: 'Details', value: 'details' },
+	{ label: 'Dependencies', value: 'dependencies' },
+	{ label: 'Inputs', value: 'inputs' }
+];
 
+const activeTab = computed( router.computedQuery( 'tab', TABS[ 0 ].value ) );
 const name = computed( () => router.query.item );
 const formattedName = computed( () => formatPath( name.value ) );
-const asset = computed( () => getAssetResource( name.value )! );
-const entrypoints = computed( () => {
-	return report.connections
-		.filter( connection => connection.kind === 'entrypoint' && connection.source === name.value )
-		.map( connection => connection.target )
-} );
-
-// Download times
-const downloadTimeOriginal = computed( () => Math.round( asset.value.uncompressed / SLOW_3G * 1000 ) );
-const downloadTimeGzip = computed( () => Math.round( asset.value.gzip / SLOW_3G * 1000 ) );
-const downloadTimeBrotli = computed( () => Math.round( asset.value.brotli / SLOW_3G * 1000 ) );
 
 // Inputs
 const inputs = computed( () => {
@@ -206,11 +56,4 @@ const inputs = computed( () => {
 		.map( chunk => chunk.name )
 		.toSorted();
 } );
-
-const dependencies = computed( () => {
-	return report.dependencies
-		.filter( dependency => dependency.paths.some( path => inputs.value.some( input => input.includes( path ) ) ) )
-		.map( dependency => dependency.name )
-		.toSorted();
-} )
 </script>
